+++ src/cron/demarchesSimplifiees.ts
@@ -21,6 +21,7 @@ import {
 } from "../services/psychologists";
 import { AdeliData } from "../types/adeli";
 import { Psychologist } from "../types/psychologist";
+import { urlExists } from "../utils/url-exists";
 
 const limit = pLimit(5);
 
@@ -109,6 +110,18 @@ export const validateDossier = async (
     errors.push(
       `Le numéro ADELI ${dossier.adeliId} ne correspond pas au département ${dossier.department}`
     );
+  }
+
+  if (dossier.website) {
+    const urlWithProtocol: string = dossier.website?.startsWith("http")
+      ? dossier.website
+      : `https://${dossier.website}`;
+    const isUrlValid = await urlExists(urlWithProtocol);
+    if (!isUrlValid) {
+      errors.push(
+        `Le site web renseigné (${dossier.website}) n'est pas valide`
+      );
+    }
   }
 
   const identifier = dossier.id?.toString() ?? dossier.email;