+++ src/pages/medecins.tsx
@@ -1,8 +1,9 @@
-import React from "react";
+import React, {useState} from "react";
 
 import HeadTag from "../components/HeadTag";
 
 const Page = () => {
+  const [video, setVideo] = useState<"hide" | "show">("hide");
   return (
     <>
       <HeadTag
@@ -31,8 +32,35 @@ const Page = () => {
               </a>
             </div>
             <div className="fr-col-12 fr-col-lg-5 align-center">
-              <img src="/images/doctor.svg" height="360" alt="" />
+              <img src="/images/doctor.svg" height="360" alt=""/>
+              {video === "hide" && (
+                <div className="align-left">
+                  <button
+                    onClick={() => setVideo("show")}
+                    title="Découvrir MonPsy en vidéo, via le témoignage de M. Werner, médecin pédiatre"
+                    className="fr-btn fr-btn--lg fr-btn--secondary fr-fi-play-line fr-btn--icon-left fr-mt-4w"
+                  >
+                    Découvrez le témoignage de M. Werner, médecin pédiatre
+                  </button>
+                  <span className="d-block fr-mt-1w">Durée&nbsp;: 2min</span>
+                </div>
+              )}
             </div>
+            {video === "show" && (
+              <div className="fr-grid-row fr-grid-row--center fr-grid-row--middle fr-pb-8w">
+                <div className="fr-col-8">
+                  <video width="100%" controls autoPlay>
+                    <source type="video/mp4" src="/images/Video-Medecin.mp4"/>
+                    <track
+                      kind="captions"
+                      srcLang="fr"
+                      src="/images/Video-Medecin.mp4.vtt"
+                      label="Activés"
+                    />
+                  </video>
+                </div>
+              </div>
+            )}
           </div>
         </div>
       </section>